Refactor: Move out BitbakeDriver from lib
Only one instance of the bitbakeDriver should exist in order to
serialize the calls to bitbake. This fixes the docker commandWrappers
getting broken pipe errors from bitbake.

Only the client will have access to it.
