+++ client/src/driver/BitbakeDriver.ts
@@ -5,8 +5,8 @@
 
 import childProcess from 'child_process'
 
-import { logger } from './utils/OutputLogger'
-import { type BitbakeSettings, loadBitbakeSettings } from './BitbakeSettings'
+import { logger } from '../lib/src/utils/OutputLogger'
+import { type BitbakeSettings, loadBitbakeSettings } from '../lib/src/BitbakeSettings'
 
 /// This class is responsible for wrapping up all bitbake classes and exposing them to the extension
 export class BitbakeDriver {
